frame_sync_bb now supports more than 1 bit per symbol.

+++ include/c4fm/frame_sync_bb.h
@@ -50,7 +50,7 @@ namespace gr {
        * class. c4fm::frame_sync_bb::make is the public interface for
        * creating new instances.
        */
-      static sptr make(int frame_size, int sync_size, int bps, int sync);
+      static sptr make(int frame_size, int sync_size, int bps, unsigned long long sync);
     };
 
   } // namespace c4fm
